Black background on all screens (#5)
